first Xilinx-like custom kernel
The zedboard kernel does not work well on the ZC702 because it uses 100% cpu
even on an idle system. (Due to dma-related kernel threads.)
The Xilinx kernel does not work well with ArchLinux because it is missing the
CGROUPS feature used by systemd.
This kernel is an attempt to improve the situation. It is similar to the
more minimalistic Xilinx configuration but includes CGROUPS for a smoother
experience with ArchLinux.